// csa/src/interfaces/card.ts
interface In_Card {
    background?: string;
    boxShadow?: string;
    borderRadius?: string;
    textSize?: string;
    width?: string;
    padding?: string;
    margin?: string;
    border?: string;
    transition?: string;
    justifyContent?: string;
    alignSelf?: string;
}

export default In_Card